Fix signed/unsigned comparison warnings
git-svn-id: svn://scribus.net/trunk/Scribus@23279 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/undostack.cpp
@@ -171,9 +171,9 @@ bool UndoStack::checkSize() {
 
 void UndoStack::clear()
 {
-	for (int i = 0; i < m_undoActions_.size(); ++i)
+	for (size_t i = 0; i < m_undoActions_.size(); ++i)
 		delete m_undoActions_[i];
-	for (int i = 0; i < m_redoActions_.size(); ++i)
+	for (size_t i = 0; i < m_redoActions_.size(); ++i)
 		delete m_redoActions_[i];
 	m_undoActions_.clear();
 	m_redoActions_.clear();
